BAH-2099 | Incorporating PR comments - move unauthorised response from utils
